test(robot): add test case Test Block Volume Should Reject Filesystem Trim
