+++ e2e/libs/keywords/volume_keywords.py
@@ -353,11 +353,11 @@ class volume_keywords:
     def activate_dr_volume(self, volume_name):
         self.volume.activate(volume_name)
 
-    def create_persistentvolume_for_volume(self, volume_name, retry=True, volume_mode="Filesystem", fsType="ext4"):
-        self.volume.create_persistentvolume(volume_name, retry, volume_mode, fsType)
+    def create_persistentvolume_for_volume(self, volume_name, retry=True, volumeMode="Filesystem", fsType="ext4"):
+        self.volume.create_persistentvolume(volume_name, retry, volumeMode, fsType)
 
-    def create_persistentvolumeclaim_for_volume(self, volume_name, retry=True, volume_mode="Filesystem"):
-        self.volume.create_persistentvolumeclaim(volume_name, retry, volume_mode)
+    def create_persistentvolumeclaim_for_volume(self, volume_name, volumeMode="Filesystem", retry=True):
+        self.volume.create_persistentvolumeclaim(volume_name, volumeMode, retry)
 
     def record_volume_replica_names(self, volume_name):
         replica_list = self.replica.get(volume_name, node_name="")
