+++ e2e/libs/persistentvolume/persistentvolume.py
@@ -16,7 +16,7 @@ class PersistentVolume():
         self.api = client.CoreV1Api()
         self.retry_count, self.retry_interval = get_retry_count_and_interval()
 
-    def create(self, name, storage, volume_mode="Filesystem", fsType="ext4"):
+    def create(self, name, storage, volumeMode="Filesystem", fsType="ext4"):
         filepath = "./templates/workload/pv.yaml"
         with open(filepath, 'r') as f:
             manifest_dict = yaml.safe_load(f)
@@ -24,7 +24,7 @@ class PersistentVolume():
             manifest_dict['metadata']['name'] = name
             manifest_dict['metadata']['labels'][LABEL_TEST] = LABEL_TEST_VALUE
             manifest_dict['spec']['capacity']['storage'] = storage
-            manifest_dict['spec']['volumeMode'] = volume_mode
+            manifest_dict['spec']['volumeMode'] = volumeMode
             manifest_dict['spec']['csi']['fsType'] = fsType
             manifest_dict['spec']['csi']['volumeHandle'] = name
