--- e2e/libs/persistentvolumeclaim/constant.py
@@ -1,2 +1,0 @@
-RETRY_COUNTS = 150
-RETRY_INTERVAL = 1
