+++ e2e/libs/persistentvolumeclaim/persistentvolumeclaim.py
@@ -1,10 +1,6 @@
 import time
 import yaml
 
-from strategy import LonghornOperationStrategy
-
-from persistentvolumeclaim.crd import CRD
-
 from kubernetes import client
 from kubernetes.client.rest import ApiException
 
@@ -18,14 +14,11 @@ from utility.utility import logging
 
 class PersistentVolumeClaim():
 
-    _strategy = LonghornOperationStrategy.CRD
-
     def __init__(self):
+        self.core_v1_api = client.CoreV1Api()
         self.retry_count, self.retry_interval = get_retry_count_and_interval()
-        if self._strategy == LonghornOperationStrategy.CRD:
-            self.claim = CRD()
 
-    def create(self, name, volume_type, sc_name, storage_size="3GiB", dataSourceName=None, dataSourceKind=None, volume_mode="Filesystem", wait_for_bound=True, volume_name=None):
+    def create(self, name, volume_type, sc_name, storage_size="3GiB", dataSourceName=None, dataSourceKind=None, volumeMode="Filesystem", wait_for_bound=True, volume_name=None):
         storage_size_bytes = convert_size_to_bytes(storage_size)
 
         filepath = "./templates/workload/pvc.yaml"
@@ -59,26 +52,23 @@ class PersistentVolumeClaim():
                 if dataSourceKind == 'VolumeSnapshot':
                     manifest_dict['spec']['dataSource']['apiGroup'] = 'snapshot.storage.k8s.io'
 
-            if volume_mode == 'Block':
-                manifest_dict['spec']['volumeMode'] = 'Block'
+            if volumeMode:
+                manifest_dict['spec']['volumeMode'] = volumeMode
 
             if volume_name:
                 manifest_dict['spec']['volumeName'] = volume_name
 
             logging(f"yaml = {manifest_dict}")
 
-            api = client.CoreV1Api()
-
-            api.create_namespaced_persistent_volume_claim(
+            self.core_v1_api.create_namespaced_persistent_volume_claim(
                 body=manifest_dict,
                 namespace=namespace)
             if wait_for_bound:
                 self.wait_for_pvc_phase(name, "Bound")
 
     def delete(self, name, namespace='default'):
-        api = client.CoreV1Api()
         try:
-            api.delete_namespaced_persistent_volume_claim(
+            self.core_v1_api.delete_namespaced_persistent_volume_claim(
                 name=name,
                 namespace=namespace,
                 grace_period_seconds=0)
@@ -95,56 +85,127 @@ class PersistentVolumeClaim():
 
     def is_exist(self, name, namespace='default'):
         exist = False
-        api = client.CoreV1Api()
-        resp = api.list_namespaced_persistent_volume_claim(namespace=namespace)
+        resp = self.core_v1_api.list_namespaced_persistent_volume_claim(namespace=namespace)
         for item in resp.items:
             if item.metadata.name == name:
                 exist = True
                 break
         return exist
 
-    def get(self, claim_name):
-        return self.claim.get(claim_name)
+    def get(self, claim_name, claim_namespace="default"):
+        return self.core_v1_api.read_namespaced_persistent_volume_claim(
+            name=claim_name,
+            namespace=claim_namespace
+        )
 
     def get_volume_name(self, claim_name):
-        api = client.CoreV1Api()
-        pvc = api.read_namespaced_persistent_volume_claim(name=claim_name, namespace='default')
+        pvc = self.core_v1_api.read_namespaced_persistent_volume_claim(name=claim_name, namespace='default')
         return pvc.spec.volume_name
 
     def list(self, claim_namespace="default", label_selector=None):
-        return self.claim.list(claim_namespace=claim_namespace,
-                             label_selector=label_selector)
+        return self.core_v1_api.list_namespaced_persistent_volume_claim(
+            namespace=claim_namespace,
+            label_selector=label_selector
+        ).items
 
     def set_label(self, claim_name, label_key, label_value, claim_namespace="default"):
-        return self.claim.set_label(claim_name, label_key, label_value, claim_namespace=claim_namespace)
+        for i in range(self.retry_count):
+            logging(f"Trying to set pvc {claim_name} label {label_key} = {label_value} ... ({i})")
+            try:
+                claim = self.get(claim_name, claim_namespace)
+
+                labels = claim.metadata.labels
+                if labels is None:
+                    labels = {}
+
+                if label_key in labels and labels[label_key] == label_value:
+                    return
+
+                labels[label_key] = label_value
+                claim.metadata.labels = labels
+                self.core_v1_api.patch_namespaced_persistent_volume_claim(
+                    name=claim_name,
+                    namespace=claim_namespace,
+                    body=claim
+                )
+            except Exception as e:
+                logging(f"Setting pvc {claim_name} label failed: {e}")
+            time.sleep(self.retry_interval)
+
+        assert False, f"Failed to set label {label_key} to {label_value} for PVC {claim_name}"
 
     def set_annotation(self, claim_name, annotation_key, annotation_value, claim_namespace="default"):
-        return self.claim.set_annotation(claim_name, annotation_key, annotation_value, claim_namespace=claim_namespace)
+        for i in range(self.retry_count):
+            logging(f"Trying to set pvc {claim_name} annotation {annotation_key} = {annotation_value} ... ({i})")
+            try:
+                claim = self.get(claim_name, claim_namespace)
 
-    def get_annotation_value(self, claim_name, annotation_key):
-        return self.claim.get_annotation_value(claim_name, annotation_key)
+                annotations = claim.metadata.annotations
+                if annotations is None:
+                    annotations = {}
 
-    def get_volume_name(self, claim_name):
-        return self.claim.get_volume_name(claim_name)
+                if annotation_key in annotations and annotations[annotation_key] == annotation_value:
+                    return
 
-    def expand(self, claim_name, size_in_byte, skip_retry=False):
-        expanded_size = self.claim.expand(claim_name, size_in_byte, skip_retry=skip_retry)
-        self.set_annotation(claim_name, ANNOT_EXPANDED_SIZE, str(expanded_size))
-        return expanded_size
+                annotations[annotation_key] = annotation_value
+                claim.metadata.annotations = annotations
+                self.core_v1_api.patch_namespaced_persistent_volume_claim(
+                    name=claim_name,
+                    namespace=claim_namespace,
+                    body=claim
+                )
+            except Exception as e:
+                logging(f"Setting pvc {claim_name} annotation failed: {e}")
+            time.sleep(self.retry_interval)
+
+        assert False, f"Failed to set annotation {annotation_key} to {annotation_value} for PVC {claim_name}"
+
+    def get_annotation_value(self, claim_name, annotation_key, claim_namespace="default"):
+        claim = self.get(claim_name, claim_namespace)
+        return claim.metadata.annotations[annotation_key]
+
+    def get_volume_name(self, claim_name, claim_namespace="default"):
+        claim = self.get(claim_name, claim_namespace)
+        return claim.spec.volume_name
+
+    def expand(self, claim_name, size_in_byte, namespace="default", skip_retry=False):
+        retry_count = 1 if skip_retry else self.retry_count
+        for i in range(retry_count):
+            logging(f"Trying to expand PVC {claim_name} to size {size_in_byte} ... ({i})")
+            try:
+                self.core_v1_api.patch_namespaced_persistent_volume_claim(
+                    name=claim_name,
+                    namespace=namespace,
+                    body={
+                        'spec': {
+                            'resources': {
+                                'requests': {
+                                    'storage': str(size_in_byte)
+                                }
+                            }
+                        }
+                    }
+                )
+                self.set_annotation(claim_name, ANNOT_EXPANDED_SIZE, str(size_in_byte))
+                return size_in_byte
+            except Exception as e:
+                logging(f"Exception when expanding PVC: {e}")
+            time.sleep(self.retry_interval)
+
+        assert False, f"Failed to expand {claim_name} size to {size_in_byte}"
 
     def expand_with_additional_bytes(self, claim_name, size_in_byte, skip_retry=False):
-        pvc = self.claim.get(claim_name)
+        pvc = self.get(claim_name)
         current_size = int(pvc.spec.resources.requests['storage'])
 
         target_size = current_size + size_in_byte
-        expanded_size = self.claim.expand(claim_name, target_size, skip_retry=skip_retry)
+        expanded_size = self.expand(claim_name, target_size, skip_retry=skip_retry)
         self.set_annotation(claim_name, ANNOT_EXPANDED_SIZE, str(expanded_size))
 
     def wait_for_pvc_phase(self, pvc_name, phase):
-        api = client.CoreV1Api()
         for i in range(self.retry_count):
             try:
-                pvc = api.read_namespaced_persistent_volume_claim(
+                pvc = self.core_v1_api.read_namespaced_persistent_volume_claim(
                     name=pvc_name, namespace='default')
                 logging(f"Waiting for pvc {pvc_name} phase to be {phase}, currently it's {pvc.status.phase} ... ({i})")
                 if pvc.status.phase == phase:
@@ -156,8 +217,7 @@ class PersistentVolumeClaim():
 
     def get_pvc_storageclass_name(self, pvc_name):
         logging(f"Reading for pvc {pvc_name} storageclass name")
-        api = client.CoreV1Api()
-        pvc = api.read_namespaced_persistent_volume_claim(
+        pvc = self.core_v1_api.read_namespaced_persistent_volume_claim(
                     name=pvc_name, namespace='default')
         logging(f"Pvc {pvc_name} is using storageclass {pvc.spec.storage_class_name }")
