+++ e2e/libs/volume/base.py
@@ -169,13 +169,13 @@ class Base(ABC):
     def activate(self, volume_name):
         return NotImplemented
 
-    def create_persistentvolume(self, volume_name, retry, volume_mode="Filesystem", fsType="ext4"):
-        logging(f'Creating PV {volume_name} with volumeMode={volume_mode}')
+    def create_persistentvolume(self, volume_name, retry, volumeMode, fsType):
+        logging(f'Creating PV {volume_name} for volume {volume_name}')
         volume = self.get(volume_name)
         volume_size = self._get_volume_size(volume)
         assert volume_size is not None, f"Cannot determine size for volume {volume_name}"
         storage = str(convert_size_to_bytes(str(volume_size)))
-        self.pv.create(volume_name, storage, volume_mode=volume_mode, fsType=fsType)
+        self.pv.create(volume_name, storage, volumeMode, fsType)
 
         if not retry:
             return
@@ -188,13 +188,13 @@ class Base(ABC):
             time.sleep(self.retry_interval)
         assert created
 
-    def create_persistentvolumeclaim(self, volume_name, retry, volume_mode="Filesystem"):
-        logging(f'Creating PVC {volume_name} with volumeMode={volume_mode}')
+    def create_persistentvolumeclaim(self, volume_name, volumeMode, retry):
+        logging(f'Creating PVC {volume_name} for volume {volume_name}')
         volume = self.get(volume_name)
         volume_size = self._get_volume_size(volume)
         assert volume_size is not None, f"Cannot determine size for volume {volume_name}"
         storage = str(convert_size_to_bytes(str(volume_size)))
-        self.pvc.create(volume_name, "RWO", "longhorn", storage_size=storage, volume_mode=volume_mode, volume_name=volume_name)
+        self.pvc.create(volume_name, "RWO", "longhorn", storage_size=storage, volume_name=volume_name, volumeMode=volumeMode)
 
         if not retry:
             return
