+++ e2e/libs/volume/volume.py
@@ -184,11 +184,11 @@ class Volume(Base):
     def activate(self, volume_name):
         return self.volume.activate(volume_name)
 
-    def create_persistentvolume(self, volume_name, retry, volume_mode="Filesystem", fsType="ext4"):
-        return self.volume.create_persistentvolume(volume_name, retry, volume_mode, fsType)
+    def create_persistentvolume(self, volume_name, retry, volumeMode, fsType):
+        return self.volume.create_persistentvolume(volume_name, retry, volumeMode, fsType)
 
-    def create_persistentvolumeclaim(self, volume_name, retry, volume_mode="Filesystem"):
-        return self.volume.create_persistentvolumeclaim(volume_name, retry, volume_mode)
+    def create_persistentvolumeclaim(self, volume_name, volumeMode, retry):
+        return self.volume.create_persistentvolumeclaim(volume_name, volumeMode, retry)
 
     def upgrade_engine_image(self, volume_name, engine_image_name):
         return self.volume.upgrade_engine_image(volume_name, engine_image_name)
